refactor: flatten modules
Move cli into app module and error into util module.
