+++ src/main.rs
@@ -1,11 +1,10 @@
+use app::Args;
+
 use clap::Parser;
-use cli::Args;
 use tracing::Level;
 use tracing_subscriber::fmt;
 
 mod app;
-mod cli;
-mod error;
 mod service;
 mod util;
 
@@ -15,7 +14,7 @@ fn main() {
     tracing::debug!("Starting app");
 
     if let Err(ref err) = args.run() {
-        error::report_exit(err, args.verbose);
+        util::error::report_exit(err, args.verbose);
     }
 
     tracing::debug!("Closing app");
