+++ src/util/mod.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod file;
 
 #[cfg(test)]
